requirements.txt
opencv-python
numpy
pyyaml
ultralytics
scipy